Use fetcher instead of revalidator
