+++ app/routes/$show.[styles.css].ts
