+++ app/routes/$show.tsx
@@ -1,10 +1,8 @@
 import '~/styles/elevation.css';
 import '~/styles/show.css';
 
-import type { LoaderFunction, MetaFunction } from '@remix-run/node';
-import { json, redirect } from '@remix-run/node';
+import type { MetaFunction } from '@remix-run/node';
 import { useLoaderData } from '@remix-run/react';
-import { addSeconds, formatISO } from 'date-fns';
 import type { FC } from 'react';
 import useLocalStorageState from 'use-local-storage-state';
 
@@ -12,9 +10,9 @@ import { AudioController } from '~/components/AudioController';
 import { ShowEnded } from '~/components/ShowEnded';
 import { ShowIntro } from '~/components/ShowIntro';
 import { ShowPlaying } from '~/components/ShowPlaying';
-import { db } from '~/db/db.server';
 import { useShowInfo } from '~/hooks/useShowInfo';
-import type { ShowData } from '~/types/ShowData';
+
+import { loader as showDataLoader } from './$show.[data.json]';
 
 export const meta: MetaFunction<typeof loader> = ({ data, params }) => {
   const id = params.show!;
@@ -26,44 +24,10 @@ export const meta: MetaFunction<typeof loader> = ({ data, params }) => {
   ];
 };
 
-export const loader = (async ({ params }) => {
-  const id = params.show!;
-
-  const show = await db.show.findUnique({
-    where: { id },
-    include: {
-      showLogoFile: true,
-      backgroundImageFile: true,
-      sets: {
-        include: {
-          audioFileUpload: { select: { audioFile: true } },
-        },
-        orderBy: { offset: 'asc' },
-      },
-    },
-  });
-  if (!show) throw redirect('/');
-
-  const data: ShowData = {
-    name: show.name,
-    description: show.description,
-    showLogoUrl: show.showLogoFile.url,
-    backgroundImageUrl: show.backgroundImageFile.url,
-    sets: show.sets.map((set) => ({
-      id: set.id,
-      audioUrl: set.audioFileUpload?.audioFile?.audioUrl ?? '',
-      artist: set.artist,
-      start: addSeconds(show.startDate, set.offset).toISOString(),
-      duration: set.audioFileUpload?.audioFile?.duration ?? 0,
-    })),
-    serverDate: formatISO(new Date()),
-  };
-
-  return json(data);
-}) satisfies LoaderFunction;
+export const loader = showDataLoader;
 
 const Show: FC = () => {
-  const loaderData = useLoaderData<typeof loader>();
+  const loaderData = useLoaderData<typeof showDataLoader>();
 
   const { targetShowInfo, onLoadedMetadata } = useShowInfo(loaderData);
